move common logic to utils
