+++ test/components/embedders/test_sentence_transformers_doc_image_embedder.py
@@ -300,81 +300,6 @@ class TestSentenceTransformersDocumentImageEmbedder:
             backend="openvino",
         )
 
-    def test_extract_image_source_info(self, test_files_path):
-        embedder = SentenceTransformersDocumentImageEmbedder(model="model")
-
-        image_paths = glob.glob(str(test_files_path / "images" / "*.*")) + glob.glob(str(test_files_path / "pdf" / "*.pdf"))
-
-        documents = []
-        for i, path in enumerate(image_paths):
-            document = Document(content=f"document number {i}", meta={"file_path": path})
-            if path.endswith(".pdf"):
-                document.meta["page_number"] = 1
-            documents.append(document)
-
-        images_source_info = embedder._extract_image_sources_info(documents=documents, file_path_meta_field="file_path", root_path="")
-        assert len(images_source_info) == len(documents)
-        for path_info in images_source_info:
-            assert str(path_info["path"]) in image_paths
-            assert path_info["type"] in ["image", "pdf"]
-            if path_info["type"] == "pdf":
-                assert path_info.get("page_number") == 1
-            else:
-                assert "page_number" not in path_info
-
-    def test_extract_image_source_info_errors(self, test_files_path):
-        embedder = SentenceTransformersDocumentImageEmbedder(model="model", file_path_meta_field="file_path")
-
-        document = Document(content="test")
-        with pytest.raises(ValueError, match="missing the 'file_path' key"):
-            embedder._extract_image_sources_info(documents=[document], file_path_meta_field="file_path", root_path="")
-
-        document = Document(content="test", meta={"file_path": "invalid_path"})
-        with pytest.raises(ValueError, match="has an invalid file path"):
-            embedder._extract_image_sources_info(documents=[document], file_path_meta_field="file_path", root_path="")
-
-        document = Document(content="test", meta={"file_path": str(test_files_path / "docx" / "sample_docx.docx")})
-        with pytest.raises(ValueError, match="has an unsupported MIME type"):
-            embedder._extract_image_sources_info(documents=[document], file_path_meta_field="file_path", root_path="")
-
-        document = Document(content="test", meta={"file_path": str(test_files_path / "pdf" / "sample_pdf_1.pdf")})
-        with pytest.raises(ValueError, match="missing the 'page_number' key"):
-            embedder._extract_image_sources_info(documents=[document], file_path_meta_field="file_path", root_path="")
-
-
-    @patch(f"{IMPORT_PATH}._convert_pdf_to_pil_images")
-    def test_process_pdf_files(self, mocked_convert_pdf_to_pil_images, test_files_path):
-        embedder = SentenceTransformersDocumentImageEmbedder(model="model")
-
-        mocked_convert_pdf_to_pil_images.return_value = [(1, Image.new("RGB", (100, 100))), (2, Image.new("RGB", (100, 100)))]
-
-        pdf_path = test_files_path / "pdf" / "sample_pdf_1.pdf"
-        pdf_doc_1: _PdfPageInfo = {"doc_idx": 0, "path": pdf_path, "page_number": 1}
-        pdf_doc_2: _PdfPageInfo = {"doc_idx": 1, "path": pdf_path, "page_number": 2}
-        pdf_documents = [pdf_doc_1, pdf_doc_2]
-
-        result = embedder._process_pdf_files(pdf_pages_info=pdf_documents)
-
-        pdf_bytestream = ByteStream.from_file_path((pdf_path))
-
-        mocked_convert_pdf_to_pil_images.assert_called_once_with(
-            bytestream=pdf_bytestream,
-            page_range=[1, 2],
-            size=None
-        )
-
-        assert len(result) == len(pdf_documents)
-        assert 0 in result and 1 in result
-        assert isinstance(result[0], Image.Image)
-        assert isinstance(result[1], Image.Image)
-
-    def test_process_pdf_files_no_pages_info(self):
-        embedder = SentenceTransformersDocumentImageEmbedder(model="model")
-        result = embedder._process_pdf_files(pdf_pages_info=[])
-
-        assert isinstance(result, dict)
-        assert len(result) == 0
-
     @pytest.mark.integration
     @pytest.mark.skipif(sys.platform == "darwin",
                         reason=("This model does not play well with GitHub macOS runners and"
